+++ lib/tools/text_tools/markdown_preview/markdown_preview_view.dart
@@ -1,7 +1,9 @@
+import 'package:alga/widgets/app_scaffold.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
 import 'package:markdown/markdown.dart' show markdownToHtml;
 
 import 'package:alga/constants/import_helper.dart';
+import 'package:url_launcher/url_launcher_string.dart';
 
 part './markdown_preview_provider.dart';
 
@@ -15,57 +17,58 @@ class MarkdownPreviewView extends StatefulWidget {
 class _MarkdownPreviewViewState extends State<MarkdownPreviewView> {
   @override
   Widget build(BuildContext context) {
-    return ToolView(
+    final bool small = isSmallDevice(context);
+    final editable = AppTitleWrapper(
+      title: S.of(context).markdownInput,
+      expand: !small,
+      actions: [
+        PasteButton(onPaste: (ref, value) {
+          ref.watch(_inputController).text = value;
+          ref.refresh(_inputValue);
+        }),
+      ],
+      child: Consumer(builder: (context, ref, _) {
+        return LangTextField(
+          expands: !small,
+          maxLines: small ? 16 : null,
+          minLines: small ? 16 : null,
+          controller: ref.watch(_inputController),
+          onChanged: (_) {
+            ref.refresh(_inputValue);
+          },
+          lang: LangHighlightType.markdown,
+        );
+      }),
+    );
+    final preview = AppTitleWrapper(
+      title: S.of(context).markdownPreviewInput,
+      expand: !small,
+      actions: [
+        CopyButton(onCopy: (ref) {
+          final data = ref.read(_inputValue);
+          return markdownToHtml(data);
+        }),
+      ],
+      child: Consumer(builder: (context, ref, _) {
+        return MarkdownBody(
+          data: ref.watch(_inputValue),
+          selectable: true,
+          onTapLink: (text, href, title) {
+            if (href != null) {
+              launchUrlString(
+                href,
+                mode: LaunchMode.externalApplication,
+              );
+            }
+          },
+        );
+      }),
+    );
+
+    return AppScaffold(
       title: Text(S.of(context).markdownPreview),
-      content: Padding(
-        padding: const EdgeInsets.all(12),
-        child: Row(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Expanded(
-              child: AppTitleWrapper(
-                title: S.of(context).markdownInput,
-                expand: true,
-                actions: [
-                  PasteButton(onPaste: (ref, value) {
-                    ref.watch(_inputController).text = value;
-                    ref.refresh(_inputValue);
-                  }),
-                ],
-                child: Consumer(builder: (context, ref, _) {
-                  return LangTextField(
-                    expands: true,
-                    controller: ref.watch(_inputController),
-                    onChanged: (_) {
-                      ref.refresh(_inputValue);
-                    },
-                    lang: LangHighlightType.markdown,
-                  );
-                }),
-              ),
-            ),
-            const SizedBox(width: 8),
-            Expanded(
-              child: AppTitleWrapper(
-                title: S.of(context).markdownPreviewInput,
-                actions: [
-                  CopyButton(onCopy: (ref) {
-                    final data = ref.read(_inputValue);
-                    return markdownToHtml(data);
-                  }),
-                ],
-                expand: true,
-                child: Consumer(builder: (context, ref, _) {
-                  return Markdown(
-                    data: ref.watch(_inputValue),
-                    selectable: true,
-                  );
-                }),
-              ),
-            ),
-          ],
-        ),
-      ),
+      first: editable,
+      second: preview,
     );
   }
 }